refactor: move prompt generator to another file
